# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Sports Fashion Store API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  category: string;
  imageUrl: string;
  /** @nullable */
  badge?: string | null;
  inStock: boolean;
  featured?: boolean;
}

export interface ProductInput {
  name: string;
  description?: string;
  price: number;
  originalPrice?: number;
  category: string;
  imageUrl: string;
  badge?: string;
  inStock?: boolean;
  featured?: boolean;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  productCount: number;
  /** @nullable */
  imageUrl?: string | null;
}

export interface SocialLink {
  id: number;
  platform: string;
  url: string;
  handle: string;
  /** @nullable */
  followerCount?: string | null;
}

export interface StoreStats {
  totalProducts: number;
  totalCategories: number;
  featuredCount: number;
  brandsCount: number;
}

export type ListProductsParams = {
category?: string;
featured?: boolean;
};