# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sports Fashion Store API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Product operations
  - name: categories
    description: Category operations
  - name: social
    description: Social/network links
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: featured
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found

  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all product categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"

  /featured:
    get:
      operationId: listFeatured
      tags: [products]
      summary: Get featured products for hero section
      responses:
        "200":
          description: Featured products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /social-links:
    get:
      operationId: listSocialLinks
      tags: [social]
      summary: Get all social/network links
      responses:
        "200":
          description: Social links
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SocialLink"

  /store-stats:
    get:
      operationId: getStoreStats
      tags: [products]
      summary: Get store statistics
      responses:
        "200":
          description: Store statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Product:
      type: object
      required:
        - id
        - name
        - price
        - category
        - imageUrl
        - inStock
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        category:
          type: string
        imageUrl:
          type: string
        badge:
          type: ["string", "null"]
        inStock:
          type: boolean
        featured:
          type: boolean

    ProductInput:
      type: object
      required:
        - name
        - price
        - category
        - imageUrl
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        originalPrice:
          type: number
        category:
          type: string
        imageUrl:
          type: string
        badge:
          type: string
        inStock:
          type: boolean
        featured:
          type: boolean

    Category:
      type: object
      required:
        - id
        - name
        - slug
        - productCount
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        productCount:
          type: integer
        imageUrl:
          type: ["string", "null"]

    SocialLink:
      type: object
      required:
        - id
        - platform
        - url
        - handle
      properties:
        id:
          type: integer
        platform:
          type: string
        url:
          type: string
        handle:
          type: string
        followerCount:
          type: ["string", "null"]

    StoreStats:
      type: object
      required:
        - totalProducts
        - totalCategories
        - featuredCount
        - brandsCount
      properties:
        totalProducts:
          type: integer
        totalCategories:
          type: integer
        featuredCount:
          type: integer
        brandsCount:
          type: integer