# lib/api-zod/src/generated/types/category.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Sports Fashion Store API
 * OpenAPI spec version: 0.1.0
 */

export interface Category {
  id: number;
  name: string;
  slug: string;
  productCount: number;
  /** @nullable */
  imageUrl?: string | null;
}