# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Sports Fashion Store API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  name: string;
  description?: string;
  price: number;
  originalPrice?: number;
  category: string;
  imageUrl: string;
  badge?: string;
  inStock?: boolean;
  featured?: boolean;
}