# lib/api-zod/src/generated/types/socialLink.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Sports Fashion Store API
 * OpenAPI spec version: 0.1.0
 */

export interface SocialLink {
  id: number;
  platform: string;
  url: string;
  handle: string;
  /** @nullable */
  followerCount?: string | null;
}