# lib/api-zod/src/generated/types/storeStats.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Sports Fashion Store API
 * OpenAPI spec version: 0.1.0
 */

export interface StoreStats {
  totalProducts: number;
  totalCategories: number;
  featuredCount: number;
  brandsCount: number;
}